fix OpenMP performance penalty via omp atomic
This commit fixes a performance penalty caused
by the use of "omp critical" that can be
trivially replaced with "omp atomic", leading to
a performance gain by about 25 percent, measured
for 1M particles on the MPCDF Raven HPC system.

Moreover, a Makefile for MPCDF Raven is contributed which
uses optimizing compiler flags, leading to a performance
benefit of similar order.