revert b6a4875 because dead loop happens
for simulations without binary this problem does not happen (e.g., some benchmark runs). However using Dragon 3 input (with hard binaries) the dead loop happens.

intgrt.f:399 GO TO 1 will loop to this line infinitely. Debug with `watch TIME` shows TIME was reset to zero in the middle of this dead loop, and statement at intgrt.F:392, IF (TIME.GT.TPREV), is never satisfied. Also witnessed that ICALL in this case increases without stop, as the dead loop goes on.

Revert this change fixed this problem.